+++ src/env/mod.rs
@@ -29,7 +29,7 @@ pub use self::fd::{FileDescEnv, FileDescEnvironment};
 pub use self::func::{FnEnv, FunctionEnvironment, UnsetFunctionEnvironment};
 pub use self::last_status::{LastStatusEnv, LastStatusEnvironment};
 pub use self::reversible_redirect::{RedirectEnvRestorer, RedirectRestorer};
-pub use self::reversible_var::{VarEnvRestorer, VarRestorer};
+pub use self::reversible_var::{VarEnvRestorer, VarEnvRestorer2, VarRestorer};
 pub use self::string_wrapper::StringWrapper;
 pub use self::var::{ExportedVariableEnvironment, VarEnv, VariableEnvironment,
                     UnsetVariableEnvironment};