+++ src/env/reversible_var.rs
@@ -13,8 +13,8 @@ pub trait VarEnvRestorer<E: ?Sized + VariableEnvironment> {
     /// Reserves capacity for at least `additional` more variables to be backed up.
     fn reserve(&mut self, additional: usize);
 
-    /// Backup and set the value of some variable, maintaining its status as an
-    /// environment variable if previously set as such.
+    /// Backup and set the value of some variable (and its exported status).
+    #[deprecated(note = "Use `VarEnvRestorer2::set_exported_var2` instead")]
     fn set_exported_var(&mut self, name: E::VarName, val: E::Var, exported: bool, env: &mut E);
 
     /// Backup and unset the value of some variable (including environment variables).
@@ -40,6 +40,7 @@ impl<'a, T, E: ?Sized> VarEnvRestorer<E> for &'a mut T
         (**self).reserve(additional);
     }
 
+    #[allow(deprecated)]
     fn set_exported_var(&mut self, name: E::VarName, val: E::Var, exported: bool, env: &mut E) {
         (**self).set_exported_var(name, val, exported, env);
     }
@@ -54,6 +55,44 @@ impl<'a, T, E: ?Sized> VarEnvRestorer<E> for &'a mut T
 
     fn restore(&mut self, env: &mut E) {
         (**self).restore(env);
+    }
+}
+
+/// A corrected extension of the `VarEnvRestorer` trait, provided in a backwards
+/// compatible manner to avoid changes on the original trait.
+///
+/// An interface for maintaining a state of all variable definitions that have
+/// been modified so that they can be restored later.
+///
+/// > *Note*: the caller should take care that a restorer instance is always
+/// > called with the same environment for its entire lifetime. Using different
+/// > environments with the same restorer instance will undoubtedly do the wrong
+/// > thing eventually, and no guarantees can be made.
+pub trait VarEnvRestorer2<E: ?Sized + VariableEnvironment>: VarEnvRestorer<E> {
+    /// Backup and set the value of some variable, either explicitly setting its
+    /// exported status as specified, or maintaining its status as an environment
+    /// variable if previously set as such.
+    fn set_exported_var2(
+        &mut self,
+        name: E::VarName,
+        val: E::Var,
+        exported: Option<bool>,
+        env: &mut E
+    );
+}
+
+impl<'a, T, E: ?Sized> VarEnvRestorer2<E> for &'a mut T
+    where T: VarEnvRestorer2<E>,
+          E: VariableEnvironment,
+{
+    fn set_exported_var2(
+        &mut self,
+        name: E::VarName,
+        val: E::Var,
+        exported: Option<bool>,
+        env: &mut E
+    ) {
+        (**self).set_exported_var2(name, val, exported, env);
     }
 }
 
@@ -155,12 +194,22 @@ impl<E: ?Sized> VarRestorer<E>
     /// environment variable if previously set as such.
     #[deprecated(note = "use the `VarEnvRestorer` trait instead")]
     pub fn set_exported_var(&mut self, name: E::VarName, val: E::Var, exported: bool, env: &mut E) {
-        self._set_exported_var(name, val, exported, env);
+        self._set_exported_var(name, val, Some(exported), env);
     }
 
-    fn _set_exported_var(&mut self, name: E::VarName, val: E::Var, exported: bool, env: &mut E) {
+    fn _set_exported_var(
+        &mut self,
+        name: E::VarName,
+        val: E::Var,
+        exported: Option<bool>,
+        env: &mut E
+    ) {
         self._backup(name.clone(), env);
-        env.set_exported_var(name, val, exported);
+
+        match exported {
+            Some(exported) => env.set_exported_var(name, val, exported),
+            None => env.set_var(name, val),
+        }
     }
 
     /// Backup and unset the value of some variable (including environment
@@ -201,7 +250,7 @@ impl<E: ?Sized> VarEnvRestorer<E> for VarRestorer<E>
     }
 
     fn set_exported_var(&mut self, name: E::VarName, val: E::Var, exported: bool, env: &mut E) {
-        self._set_exported_var(name, val, exported, env);
+        self._set_exported_var(name, val, Some(exported), env);
     }
 
     fn unset_var(&mut self, name: E::VarName, env: &mut E) {
@@ -215,5 +264,21 @@ impl<E: ?Sized> VarEnvRestorer<E> for VarRestorer<E>
 
     fn restore(&mut self, env: &mut E) {
         self._restore(env);
+    }
+}
+
+impl<E: ?Sized> VarEnvRestorer2<E> for VarRestorer<E>
+    where E: ExportedVariableEnvironment + UnsetVariableEnvironment,
+          E::VarName: Clone,
+          E::Var: Clone,
+{
+    fn set_exported_var2(
+        &mut self,
+        name: E::VarName,
+        val: E::Var,
+        exported: Option<bool>,
+        env: &mut E
+    ) {
+        self._set_exported_var(name, val, exported, env);
     }
 }